Address most implicit 64-to-32-bit conversion warnings.
Visual Studio and Xcode both issue such warnings by default, and Clang
also offers a -Wshorten-64-to-32 option to expose them elsewhere.  To
address them, harmonize types where practical and add casts as
appropriate (first formally capping executables' read request sizes as
necessary).  NB: detect_repeat in pcre2_jit_compile.c still yields two
such warnings that look plausibly legitimate but not so easy to fix. :-/

+++ src/pcre2_internal.h
@@ -2116,7 +2116,7 @@ extern void *       _pcre2_memmove(void *, const void *, size_t);
 
 #endif  /* PCRE2_CODE_UNIT_WIDTH */
 
-extern BOOL         PRIV(ckd_smul)(PCRE2_SIZE *, int, int);
+extern BOOL         PRIV(ckd_smul)(PCRE2_SIZE *, ptrdiff_t, ptrdiff_t);
 
 #include "pcre2_util.h"
